+++ src/mfx/uitk/pg/ParamEdit.cpp
@@ -68,7 +68,7 @@ ParamEdit::ParamEdit (PageSwitcher &page_switcher, LocEdit &loc_edit)
 ,	_step_index (0)
 {
 	std::string    ratio;
-	for (size_t k = 0; k < _nbr_steps; ++k)
+	for (int k = 0; k < _nbr_steps; ++k)
 	{
 		ratio += "\xE2\x9A\xAB";   // MEDIUM BLACK CIRCLE U+26AB
 		_step_sptr_arr [k] = TxtSPtr (new NText (Entry_STEP + k));
